Improve Kotlin template
Put the rule outside of method scope so that it won't recreate the rule everytime it
does conversion process.

Use object instead of companion object, making it build only one class file instead of
two.

Add JvmStatic annotation to method names for interop with Java. It allows the methods
to be called directly from Java instead of through an instance